ai-service/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pydantic